refactor: move task pkg to common
