+++ ghx/step_action.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/aweris/gale/common/log"
 	"github.com/aweris/gale/common/model"
+	"github.com/aweris/gale/common/task"
 	"github.com/aweris/gale/ghx/context"
-	"github.com/aweris/gale/ghx/task"
 )
 
 var (
